converted IRTFileImport.tsx to typescript
