+++ services/minespace-web/src/tests/components/project/informationRequirementsTablePage/InformationRequirementsTablePage.spec.tsx
